requirements.txt
beautifulsoup4==4.12.3
requests==2.32.2
pandas-2.2.3
xlsxwriter-3.2.0
selenium-4.26.1
